lesson2_langchain_core/requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
python-dotenv
streamlit
groq
openai